+++ src/main/scala/uk/gov/hmrc/govukfrontend/views/Implicits.scala
@@ -205,6 +205,7 @@ trait Implicits {
       * Extension method to allow a Play form Field to be used to add certain parameters in a Radios,
       * specifically errorMessage, idPrefix, name, and checked (for a specific RadioItem). Note these
       * values will only be added from the Field if they are not specifically defined in the Radios object.
+      * Form errors will be bound as Text objects.
       *
       * @param field
       * @param messages
@@ -214,7 +215,21 @@ trait Implicits {
         .withName(field)
         .withIdPrefix(field)
         .withItemsChecked(field)
-        .withErrorMessage(field)
+        .withTextErrorMessage(field)
+
+    /**
+      * Extension method to allow a Play form Field to be used to add certain parameters in a Radios, as per
+      * withFormField, with form errors bound as HtmlContent objects.
+      *
+      * @param field
+      * @param messages
+      */
+    override def withFormFieldWithErrorAsHtml(field: Field): Radios =
+      radios
+        .withName(field)
+        .withIdPrefix(field)
+        .withItemsChecked(field)
+        .withHtmlErrorMessage(field)
 
     private[views] def withName(field: Field): Radios =
       withStringProperty(field.name, radios.name, radios)((rds, nm) => rds.copy(name = nm))
@@ -222,8 +237,13 @@ trait Implicits {
     private[views] def withIdPrefix(field: Field): Radios =
       withOptStringProperty(Some(field.name), radios.idPrefix, radios)((rds, ip) => rds.copy(idPrefix = ip))
 
-    private[views] def withErrorMessage(field: Field): Radios =
-      withOptErrorMessageProperty(field.error, radios.errorMessage, radios)((rds, errorMsg) =>
+    private[views] def withTextErrorMessage(field: Field): Radios =
+      withOptTextErrorMessageProperty(field.error, radios.errorMessage, radios)((rds, errorMsg) =>
+        rds.copy(errorMessage = errorMsg)
+      )
+
+    private[views] def withHtmlErrorMessage(field: Field): Radios =
+      withOptHtmlErrorMessageProperty(field.error, radios.errorMessage, radios)((rds, errorMsg) =>
         rds.copy(errorMessage = errorMsg)
       )
 
@@ -241,9 +261,10 @@ trait Implicits {
   implicit class RichInput(input: Input)(implicit val messages: Messages) extends ImplicitsSupport[Input] {
 
     /**
-      * Extension method to allow a Play form Field to be used to add certain parameters in an Input,
-      * specifically errorMessage, id, name, and value. Note these
-      * values will only be added from the Field if they are not specifically defined in the Input object.
+      * Extension method to allow a Play form Field to be used to add certain parameters in an Input, specifically
+      * errorMessage, id, name, and value. Note these values will only be added from the Field if they are not
+      * specifically defined in the Input object.
+      * Form errors will be bound as Text objects.
       *
       * @param field
       * @param messages
@@ -253,7 +274,22 @@ trait Implicits {
         .withName(field)
         .withId(field)
         .withValue(field)
-        .withErrorMessage(field)
+        .withTextErrorMessage(field)
+
+    /**
+      * Extension method to allow a Play form Field to be used to add certain parameters in an Input, as per
+      * withFormField, with form errors bound as HtmlContent objects.
+      *
+      * @param field
+      * @param messages
+      */
+
+    override def withFormFieldWithErrorAsHtml(field: Field): Input =
+      input
+        .withName(field)
+        .withId(field)
+        .withValue(field)
+        .withHtmlErrorMessage(field)
 
     private[views] def withName(field: Field): Input =
       withStringProperty(field.name, input.name, input)((ipt, nm) => ipt.copy(name = nm))
@@ -264,8 +300,13 @@ trait Implicits {
     private[views] def withValue(field: Field): Input =
       withOptStringProperty(field.value, input.value, input)((ipt, vl) => ipt.copy(value = vl))
 
-    private[views] def withErrorMessage(field: Field): Input =
-      withOptErrorMessageProperty(field.error, input.errorMessage, input)((ipt, errorMsg) =>
+    private[views] def withTextErrorMessage(field: Field): Input =
+      withOptTextErrorMessageProperty(field.error, input.errorMessage, input)((ipt, errorMsg) =>
+        ipt.copy(errorMessage = errorMsg)
+      )
+
+    private[views] def withHtmlErrorMessage(field: Field): Input =
+      withOptHtmlErrorMessageProperty(field.error, input.errorMessage, input)((ipt, errorMsg) =>
         ipt.copy(errorMessage = errorMsg)
       )
   }
@@ -277,6 +318,7 @@ trait Implicits {
       * Extension method to allow a Play form Field to be used to add certain parameters in a Checkboxes,
       * specifically errorMessage, idPrefix, name, and checked (for a specific CheckboxItem). Note these
       * values will only be added from the Field if they are not specifically defined in the Checkboxes object.
+      * Form errors will be bound as Text objects.
       *
       * @param field
       * @param messages
@@ -285,7 +327,22 @@ trait Implicits {
       checkboxes
         .withName(field)
         .withIdPrefix(field)
-        .withErrorMessage(field)
+        .withTextErrorMessage(field)
+        .withItemsChecked(field)
+
+    /**
+      * Extension method to allow a Play form Field to be used to add certain parameters in a Checkboxes, as per
+      * withFormField, with form errors bound as HtmlContent objects.
+      *
+      * @param field
+      * @param messages
+      */
+
+    override def withFormFieldWithErrorAsHtml(field: Field): Checkboxes =
+      checkboxes
+        .withName(field)
+        .withIdPrefix(field)
+        .withHtmlErrorMessage(field)
         .withItemsChecked(field)
 
     private[views] def withName(field: Field): Checkboxes =
@@ -294,8 +351,13 @@ trait Implicits {
     private[views] def withIdPrefix(field: Field): Checkboxes =
       withOptStringProperty(Some(field.name), checkboxes.idPrefix, checkboxes)((cb, ip) => cb.copy(idPrefix = ip))
 
-    private[views] def withErrorMessage(field: Field): Checkboxes =
-      withOptErrorMessageProperty(field.error, checkboxes.errorMessage, checkboxes)((cb, errorMsg) =>
+    private[views] def withTextErrorMessage(field: Field): Checkboxes =
+      withOptTextErrorMessageProperty(field.error, checkboxes.errorMessage, checkboxes)((cb, errorMsg) =>
+        cb.copy(errorMessage = errorMsg)
+      )
+
+    private[views] def withHtmlErrorMessage(field: Field): Checkboxes =
+      withOptHtmlErrorMessageProperty(field.error, checkboxes.errorMessage, checkboxes)((cb, errorMsg) =>
         cb.copy(errorMessage = errorMsg)
       )
 
@@ -315,9 +377,10 @@ trait Implicits {
   implicit class RichSelect(select: Select)(implicit val messages: Messages) extends ImplicitsSupport[Select] {
 
     /**
-      * Extension method to allow a Play form Field to be used to add certain parameters in a Select,
-      * specifically errorMessage, id, name, and selected (for a specific SelectItem). Note these
-      * values will only be added from the Field if they are not specifically defined in the Select object.
+      * Extension method to allow a Play form Field to be used to add certain parameters in a Select, specifically
+      * errorMessage, id, name, and selected (for a specific SelectItem). Note these values will only be added from the
+      * Field if they are not specifically defined in the Select object.
+      * Form errors will be bound as Text objects.
       *
       * @param field
       * @param messages
@@ -326,7 +389,21 @@ trait Implicits {
       select
         .withName(field)
         .withId(field)
-        .withErrorMessage(field)
+        .withTextErrorMessage(field)
+        .withItemSelected(field)
+
+    /**
+      * Extension method to allow a Play form Field to be used to add certain parameters in a Select, as per
+      * withFormField, with form errors bound as HtmlContent objects.
+      *
+      * @param field
+      * @param messages
+      */
+    override def withFormFieldWithErrorAsHtml(field: Field): Select =
+      select
+        .withName(field)
+        .withId(field)
+        .withHtmlErrorMessage(field)
         .withItemSelected(field)
 
     private[views] def withName(field: Field): Select =
@@ -335,8 +412,13 @@ trait Implicits {
     private[views] def withId(field: Field): Select =
       withStringProperty(field.name, select.id, select)((sct, id) => sct.copy(id = id))
 
-    private[views] def withErrorMessage(field: Field): Select =
-      withOptErrorMessageProperty(field.error, select.errorMessage, select)((sct, errorMsg) =>
+    private[views] def withTextErrorMessage(field: Field): Select =
+      withOptTextErrorMessageProperty(field.error, select.errorMessage, select)((sct, errorMsg) =>
+        sct.copy(errorMessage = errorMsg)
+      )
+
+    private[views] def withHtmlErrorMessage(field: Field): Select =
+      withOptHtmlErrorMessageProperty(field.error, select.errorMessage, select)((sct, errorMsg) =>
         sct.copy(errorMessage = errorMsg)
       )
 
@@ -354,9 +436,10 @@ trait Implicits {
   implicit class RichTextarea(textArea: Textarea)(implicit val messages: Messages) extends ImplicitsSupport[Textarea] {
 
     /**
-      * Extension method to allow a Play form Field to be used to add certain parameters in an Textarea,
-      * specifically errorMessage, id, name, and value. Note these
-      * values will only be added from the Field if they are not specifically defined in the Textarea object.
+      * Extension method to allow a Play form Field to be used to add certain parameters in an Textarea, specifically
+      * errorMessage, id, name, and value. Note these values will only be added from the Field if they are not
+      * specifically defined in the Textarea object.
+      * Form errors will be bound as Text objects.
       *
       * @param field
       * @param messages
@@ -366,7 +449,21 @@ trait Implicits {
         .withName(field)
         .withId(field)
         .withValue(field)
-        .withErrorMessage(field)
+        .withTextErrorMessage(field)
+
+    /**
+      * Extension method to allow a Play form Field to be used to add certain parameters in an Textarea, as per
+      * withFormField, with form errors bound as HtmlContent objects.
+      *
+      * @param field
+      * @param messages
+      */
+    override def withFormFieldWithErrorAsHtml(field: Field): Textarea =
+      textArea
+        .withName(field)
+        .withId(field)
+        .withValue(field)
+        .withHtmlErrorMessage(field)
 
     private[views] def withName(field: Field): Textarea =
       withStringProperty(field.name, textArea.name, textArea)((ta, nm) => ta.copy(name = nm))
@@ -377,8 +474,13 @@ trait Implicits {
     private[views] def withValue(field: Field): Textarea =
       withOptStringProperty(field.value, textArea.value, textArea)((ta, vl) => ta.copy(value = vl))
 
-    private[views] def withErrorMessage(field: Field): Textarea =
-      withOptErrorMessageProperty(field.error, textArea.errorMessage, textArea)((ta, errorMsg) =>
+    private[views] def withTextErrorMessage(field: Field): Textarea =
+      withOptTextErrorMessageProperty(field.error, textArea.errorMessage, textArea)((ta, errorMsg) =>
+        ta.copy(errorMessage = errorMsg)
+      )
+
+    private[views] def withHtmlErrorMessage(field: Field): Textarea =
+      withOptHtmlErrorMessageProperty(field.error, textArea.errorMessage, textArea)((ta, errorMsg) =>
         ta.copy(errorMessage = errorMsg)
       )
   }
@@ -388,8 +490,9 @@ trait Implicits {
 
     /**
       * Extension method to allow a Play form Field to be used to add certain parameters in an CharacterCount,
-      * specifically errorMessage, id, name, and value. Note these
-      * values will only be added from the Field if they are not specifically defined in the CharacterCount object.
+      * specifically errorMessage, id, name, and value. Note these values will only be added from the Field if they are
+      * not specifically defined in the CharacterCount object.
+      * Form errors will be bound as Text objects.
       *
       * @param field
       * @param messages
@@ -399,7 +502,21 @@ trait Implicits {
         .withName(field)
         .withId(field)
         .withValue(field)
-        .withErrorMessage(field)
+        .withTextErrorMessage(field)
+
+    /**
+      * Extension method to allow a Play form Field to be used to add certain parameters in an CharacterCount, as per
+      * withFormField, with form errors bound as HtmlContent objects.
+      *
+      * @param field
+      * @param messages
+      */
+    override def withFormFieldWithErrorAsHtml(field: Field): CharacterCount =
+      characterCount
+        .withName(field)
+        .withId(field)
+        .withValue(field)
+        .withHtmlErrorMessage(field)
 
     private[views] def withName(field: Field): CharacterCount =
       withStringProperty(field.name, characterCount.name, characterCount)((cc, nm) => cc.copy(name = nm))
@@ -410,8 +527,13 @@ trait Implicits {
     private[views] def withValue(field: Field): CharacterCount =
       withOptStringProperty(field.value, characterCount.value, characterCount)((cc, vl) => cc.copy(value = vl))
 
-    private[views] def withErrorMessage(field: Field): CharacterCount =
-      withOptErrorMessageProperty(field.error, characterCount.errorMessage, characterCount)((cc, errorMsg) =>
+    private[views] def withTextErrorMessage(field: Field): CharacterCount =
+      withOptTextErrorMessageProperty(field.error, characterCount.errorMessage, characterCount)((cc, errorMsg) =>
+        cc.copy(errorMessage = errorMsg)
+      )
+
+    private[views] def withHtmlErrorMessage(field: Field): CharacterCount =
+      withOptHtmlErrorMessageProperty(field.error, characterCount.errorMessage, characterCount)((cc, errorMsg) =>
         cc.copy(errorMessage = errorMsg)
       )
   }
@@ -425,6 +547,8 @@ trait ImplicitsSupport[T] {
   implicit val messages: Messages
 
   def withFormField(field: Field): T
+
+  def withFormFieldWithErrorAsHtml(field: Field): T
 
   protected[views] def withStringProperty(propFromField: String, currentProp: String, currentFormInput: T)(
     update: (T, String) => T
@@ -448,13 +572,25 @@ trait ImplicitsSupport[T] {
       formInput = currentFormInput
     )(update)
 
-  protected[views] def withOptErrorMessageProperty(
+  protected[views] def withOptTextErrorMessageProperty(
     formError: Option[FormError],
     currentProp: Option[ErrorMessage],
     currentFormInput: T
   )(update: (T, Option[ErrorMessage]) => T): T =
     withProperty[Option[ErrorMessage], T](
-      propertyFromField = formErrorToErrorMessage(formError),
+      propertyFromField = formErrorToTextErrorMessage(formError),
+      propertyFromUnderlying = currentProp,
+      default = None,
+      formInput = currentFormInput
+    )(update)
+
+  protected[views] def withOptHtmlErrorMessageProperty(
+    formError: Option[FormError],
+    currentProp: Option[ErrorMessage],
+    currentFormInput: T
+  )(update: (T, Option[ErrorMessage]) => T): T =
+    withProperty[Option[ErrorMessage], T](
+      propertyFromField = formErrorToHtmlErrorMessage(formError),
       propertyFromUnderlying = currentProp,
       default = None,
       formInput = currentFormInput
@@ -466,6 +602,9 @@ trait ImplicitsSupport[T] {
     if (propertyFromUnderlying == default) update(formInput, propertyFromField)
     else formInput
 
-  private def formErrorToErrorMessage(formError: Option[FormError]): Option[ErrorMessage] =
+  private def formErrorToTextErrorMessage(formError: Option[FormError]): Option[ErrorMessage] =
     formError.map(formError => ErrorMessage(content = Text(messages(formError.message, formError.args: _*))))
+
+  private def formErrorToHtmlErrorMessage(formError: Option[FormError]): Option[ErrorMessage] =
+    formError.map(formError => ErrorMessage(content = HtmlContent(messages(formError.message, formError.args: _*))))
 }